+++ src/main/concrete_ingredients/Pancake.java
@@ -5,8 +5,8 @@ import main.interfaces.IVisitor;
 
 public class Pancake implements IIngredient {
     @Override public String getName() { return "Pancake"; }
-    @Override public double getCalories() { return 300; }
-    @Override public double getPrice() { return 2.5; }
+    @Override public float getCalories() { return 300; }
+    @Override public float getPrice() { return 2.5F; }
 
     @Override
     public void accept(IVisitor visitor) {
