+++ src/main/visitors/PriceVisitor.java
@@ -7,16 +7,16 @@ import main.concrete_ingredients.Syrup;
 import main.interfaces.IVisitor;
 
 public class PriceVisitor implements IVisitor {
-    private double totalPrice = 0;
+    private float totalPrice = 0;
 
     @Override public void visitPancake(Pancake p) { totalPrice += p.getPrice(); }
-    @Override public void visitSyrup(Syrup s) { totalPrice += s.getPrice(); }
+    @Override public void visitSyrup(Syrup s) { totalPrice += (float) s.getPrice(); }
     @Override public void visitEgg(Egg e) { totalPrice += e.getPrice(); }
     @Override public void visitMenu(Menu m) {
         System.out.println("Calculating price for: " + m.getName());
     }
 
-    public double getTotalPrice() {
+    public float getTotalPrice() {
         return totalPrice;
     }
 }
